requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
jsonify==0.5
MarkupSafe==2.1.5
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
